# specs/001-school-library-system/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: BCD Library Management System API
  description: |
    REST API for school library management (Bibliothèque Centre Documentaire).

    **Operational Model**: Librarian-operated system with offline-first design.
    - No fines/fees (privilege blocking instead)
    - No email/SMS notifications (printed reports)
    - Librarian-mediated access (no student self-service)

    **Key Features**:
    - Circulation (checkout/return/renew with barcode scanning)
    - Cataloging (BNF SRU API integration for ISBN lookup)
    - Borrower management (students/teachers/staff organized by class)
    - Holds/reservations (librarian-mediated queue management)
    - Reports (overdue, statistics, never-borrowed)
    - Barcode generation (Code 39/128)
  version: 1.0.0
  contact:
    name: BCD API Support
    email: support@bcd-library.local
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server (localhost)
  - url: http://library-server.local:8000/api/v1
    description: Production server (local network)
...
paths:
...
  /admin/settings:
    get:
      tags: [admin]
      summary: Get system settings
      operationId: getSettings
      responses:
        '200':
          description: System settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'

    put:
      tags: [admin]
      summary: Update system settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemSettings'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemSettings'
  /admin/backup:
    post:
      tags: [admin]
      summary: Create database backup
      operationId: createBackup
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [sql, sqlite]
            default: sqlite
      responses:
        '200':
          description: Backup file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
...
components:
  schemas:
...
    SystemSettings:
      type: object
      properties:
        id_format:
          type: string
          enum: [numeric, alphanumeric]
        id_validation_regex:
          type: string
        barcode_type:
          type: string
          enum: [code39, code128]
        loan_limit_default:
          type: integer
        loan_limit_teacher:
          type: integer
        loan_duration_days:
          type: integer
        renewal_limit:
          type: integer
        hold_expiration_days:
          type: integer
        language:
          type: string
          enum: [fr, en]
        academic_year_current:
          type: string
        library_name:
          type: string
...